+++ gusto_embedded/src/funcs/companyBenefitsPutV1CompanyBenefitsCompanyBenefitIdContributionExclusions.ts
@@ -109,7 +109,6 @@ async function $do(
       { explode: false, charEncoding: "percent" },
     ),
   };
-
   const path = pathToFunc(
     "/v1/company_benefits/{company_benefit_id}/contribution_exclusions",
   )(pathParams);