+++ gusto_embedded/src/funcs/companyBenefitsUpdate.ts
@@ -110,7 +110,6 @@ async function $do(
       { explode: false, charEncoding: "percent" },
     ),
   };
-
   const path = pathToFunc("/v1/company_benefits/{company_benefit_id}")(
     pathParams,
   );