+++ gusto_embedded/src/funcs/employeeEmploymentsGetHistory.ts
@@ -107,7 +107,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc("/v1/employees/{employee_id}/employment_history")(
     pathParams,
   );