+++ gusto_embedded/src/funcs/employeesUpdate.ts
@@ -110,7 +110,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc("/v1/employees/{employee_id}")(pathParams);
 
   const headers = new Headers(compactMap({