+++ gusto_embedded/src/funcs/jobsAndCompensationsDeleteCompensation.ts
@@ -114,7 +114,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc("/v1/compensations/{compensation_id}")(pathParams);
 
   const headers = new Headers(compactMap({