+++ gusto_embedded/src/funcs/jobsAndCompensationsUpdateCompensation.ts
@@ -118,7 +118,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc("/v1/compensations/{compensation_id}")(pathParams);
 
   const headers = new Headers(compactMap({