+++ gusto_embedded/src/funcs/locationsUpdate.ts
@@ -110,7 +110,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc("/v1/locations/{location_id}")(pathParams);
 
   const headers = new Headers(compactMap({