+++ gusto_embedded/src/funcs/payrollsSkip.ts
@@ -107,7 +107,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc("/v1/companies/{company_uuid}/payrolls/skip")(
     pathParams,
   );