+++ gusto_embedded/src/funcs/peopleBatchesPostV1CompaniesCompanyIdPeopleBatches.ts
@@ -119,7 +119,6 @@ async function $do(
       charEncoding: "percent",
     }),
   };
-
   const path = pathToFunc("/v1/companies/{company_id}/people_batches")(
     pathParams,
   );