+++ gusto_embedded/src/funcs/timeOffPoliciesAddEmployees.ts
@@ -113,7 +113,6 @@ async function $do(
       { explode: false, charEncoding: "percent" },
     ),
   };
-
   const path = pathToFunc(
     "/v1/time_off_policies/{time_off_policy_uuid}/add_employees",
   )(pathParams);