+++ gusto_embedded/src/funcs/webhooksUpdateSubscription.ts
@@ -113,7 +113,6 @@ async function $do(
       { explode: false, charEncoding: "percent" },
     ),
   };
-
   const path = pathToFunc(
     "/v1/webhook_subscriptions/{webhook_subscription_uuid}",
   )(pathParams);