+++ gusto_embedded/src/lib/config.ts
@@ -64,8 +64,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2025-06-15",
-  sdkVersion: "0.12.4",
-  genVersion: "2.866.0",
+  sdkVersion: "0.12.5",
+  genVersion: "2.869.25",
   userAgent:
-    "speakeasy-sdk/typescript 0.12.4 2.866.0 2025-06-15 @gusto/embedded-api",
+    "speakeasy-sdk/typescript 0.12.5 2.869.25 2025-06-15 @gusto/embedded-api",
 } as const;