+++ gusto_embedded/src/models/operations/getv1companiescompanyidfederaltaxdetails.ts
@@ -5,6 +5,7 @@
 import * as z from "zod/v3";
 import { remap as remap$ } from "../../lib/primitives.js";
 import { safeParse } from "../../lib/schemas.js";
+import { ClosedEnum } from "../../types/enums.js";
 import { Result as SafeParseResult } from "../../types/fp.js";
 import {
   FederalTaxDetails,
@@ -14,11 +15,21 @@ import {
   HTTPMetadata,
   HTTPMetadata$inboundSchema,
 } from "../components/httpmetadata.js";
-import {
-  VersionHeader,
-  VersionHeader$outboundSchema,
-} from "../components/versionheader.js";
 import { SDKValidationError } from "../errors/sdkvalidationerror.js";
+
+/**
+ * Determines the date-based API version associated with your API call. If none is provided, your application's [minimum API version](https://docs.gusto.com/embedded-payroll/docs/api-versioning#minimum-api-version) is used.
+ */
+export const GetV1CompaniesCompanyIdFederalTaxDetailsHeaderXGustoAPIVersion = {
+  TwoThousandAndTwentyFiveMinus06Minus15: "2025-06-15",
+} as const;
+/**
+ * Determines the date-based API version associated with your API call. If none is provided, your application's [minimum API version](https://docs.gusto.com/embedded-payroll/docs/api-versioning#minimum-api-version) is used.
+ */
+export type GetV1CompaniesCompanyIdFederalTaxDetailsHeaderXGustoAPIVersion =
+  ClosedEnum<
+    typeof GetV1CompaniesCompanyIdFederalTaxDetailsHeaderXGustoAPIVersion
+  >;
 
 export type GetV1CompaniesCompanyIdFederalTaxDetailsRequest = {
   /**
@@ -28,16 +39,26 @@ export type GetV1CompaniesCompanyIdFederalTaxDetailsRequest = {
   /**
    * Determines the date-based API version associated with your API call. If none is provided, your application's [minimum API version](https://docs.gusto.com/embedded-payroll/docs/api-versioning#minimum-api-version) is used.
    */
-  xGustoAPIVersion?: VersionHeader | undefined;
+  xGustoAPIVersion?:
+    | GetV1CompaniesCompanyIdFederalTaxDetailsHeaderXGustoAPIVersion
+    | undefined;
 };
 
 export type GetV1CompaniesCompanyIdFederalTaxDetailsResponse = {
   httpMeta: HTTPMetadata;
   /**
-   * Example response
+   * Success
    */
   federalTaxDetails?: FederalTaxDetails | undefined;
 };
+
+/** @internal */
+export const GetV1CompaniesCompanyIdFederalTaxDetailsHeaderXGustoAPIVersion$outboundSchema:
+  z.ZodNativeEnum<
+    typeof GetV1CompaniesCompanyIdFederalTaxDetailsHeaderXGustoAPIVersion
+  > = z.nativeEnum(
+    GetV1CompaniesCompanyIdFederalTaxDetailsHeaderXGustoAPIVersion,
+  );
 
 /** @internal */
 export type GetV1CompaniesCompanyIdFederalTaxDetailsRequest$Outbound = {
@@ -53,7 +74,9 @@ export const GetV1CompaniesCompanyIdFederalTaxDetailsRequest$outboundSchema:
     GetV1CompaniesCompanyIdFederalTaxDetailsRequest
   > = z.object({
     companyId: z.string(),
-    xGustoAPIVersion: VersionHeader$outboundSchema.default("2025-06-15"),
+    xGustoAPIVersion:
+      GetV1CompaniesCompanyIdFederalTaxDetailsHeaderXGustoAPIVersion$outboundSchema
+        .default("2025-06-15"),
   }).transform((v) => {
     return remap$(v, {
       companyId: "company_id",